requirements.txt
SQLAlchemy == 2.0.21
pandas == 2.0.1
loguru == 0.7.2
PyYAML == 6.0.1
numpy == 1.26.1
PyMySQL == 1.1.1
Flask == 3.1.0
requests == 2.32.3